requirements.txt
pyyaml>=3.13
RPi.GPIO>=0.7.0
astral>=2.2
websockets>=8.1
